+++ src/waterfall.h
@@ -24,5 +24,5 @@ void waterfall_band_set();
 
 void waterfall_change_max(int16_t d);
 void waterfall_change_min(int16_t d);
-void waterfall_change_freq(int16_t df);
+void waterfall_change_freq(int64_t df);
 void waterfall_update_band(uint64_t f);
